Update comments, javadocs, cleanup formatting
