+++ src/main/java/gov/usgs/earthquake/aws/JsonNotification.java
@@ -13,41 +13,58 @@ import gov.usgs.earthquake.product.Product;
 import gov.usgs.earthquake.product.io.JsonProduct;
 import gov.usgs.util.protocolhandlers.data.Handler;
 
+/**
+ * Json formatted notification.
+ *
+ * Stores stores product content in URLNotification as a data URL.
+ */
 public class JsonNotification extends URLNotification {
-	public static final URL EMPTY_URL;
-	static {
-		try {
-			// make sure data protocol handler is registered
-			Handler.register();
-			EMPTY_URL = new URL("data:,");
-		} catch (MalformedURLException mue) {
-			throw new RuntimeException("failed to parse empty url");
-		}
-	}
 
-	public final Instant created;
-	public final Product product;
+  /** Empty URL for product tracker. */
+  public static final URL EMPTY_URL;
+  static {
+    try {
+      // make sure data protocol handler is registered
+      Handler.register();
+      EMPTY_URL = new URL("data:,");
+    } catch (MalformedURLException mue) {
+      throw new RuntimeException("failed to parse empty url");
+    }
+  }
 
-	JsonNotification(final JsonObject json) throws Exception {
-		this(
-				Instant.parse(json.getString("created")),
-				new JsonProduct().getProduct(
-						json.getJsonObject("product")));
-	}
+  /** When notification was created. */
+  public final Instant created;
+  /** Product attached to notification. */
+  public final Product product;
 
-	JsonNotification(final Instant created, final Product product) throws Exception {
-		super(
-				product.getId(),
-				// expiration date
-				new Date(created.plusSeconds(30 * 86400).toEpochMilli()),
-				// no tracker
-				EMPTY_URL,
-				// store product as data url
-				new URL("data:;base64," +
-						new String(Base64.getEncoder().encode(
-								new JsonProduct().getJsonObject((product))
-										.toString().getBytes("utf8")))));
-		this.created = created;
-		this.product = product;
-	}
+  /**
+   * Parse a Json formatted Notification.
+   *
+   * @param json
+   * @throws Exception
+   */
+  JsonNotification(final JsonObject json) throws Exception {
+    this(
+        Instant.parse(json.getString("created")),
+        new JsonProduct().getProduct(json.getJsonObject("product")));
+  }
+
+  /**
+   * Create a JsonNotification from an existing Product.
+   */
+  JsonNotification(final Instant created, final Product product) throws Exception {
+    super(
+        product.getId(),
+        // expiration date
+        new Date(created.plusSeconds(30 * 86400).toEpochMilli()),
+        // no tracker
+        EMPTY_URL,
+        // store product as data url
+        new URL("data:;base64," +
+            new String(Base64.getEncoder().encode(
+                new JsonProduct().getJsonObject((product))
+                    .toString().getBytes("utf8")))));
+    this.created = created;
+    this.product = product;
+  }
 }
